+++ src/buylog/__init__.py
@@ -1,2 +1,4 @@
 from .cli import main as main
 from . import models as models
+
+__version__ = "0.1.8"